Create Remove-RemoteUserProfile.ps1
Load admin credentials to be used to connect to remote host.  

Ask and verify for the remote host name.  If the connection fails, notify the user the computer is unreachable and exit.

If reachable, the pull a listing of all profiles and create a custom object (UserName, EntryNumber and ProfilePath)  These will be used to prompt the user to select which profile(s) they wish to delete.